% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfdcs2raven.R
\name{lfdcs_to_raven}
\alias{lfdcs_to_raven}
\title{lfdcs_to_raven}
\usage{
lfdcs_to_raven(
  lfdcs_file,
  raven_file = "selections.txt",
  audio_start_time,
  lfdcs_start_time = "1970-01-01 00:00:00",
  call_types = "",
  max_mdist = "",
  t_buff = 0.5,
  f_buff = 25
)
}
\arguments{
\item{lfdcs_file}{The autodetections generated by \code{export_autodetections}}

\item{raven_file}{The desired name (.txt) of the Raven selection table}

\item{audio_start_time}{String giving the audio start time (YYYY-mm-dd HH:MM:SS)}

\item{lfdcs_start_time}{String giving the lfdcs start time (YYYY-mm-dd HH:MM:SS)}

\item{call_types}{A vector of desired call types}

\item{max_mdist}{A numeric value indicating the maximum Mahalanobis distance to export}

\item{t_buff}{The time buffer (in seconds) that is added to the start and end of each selection. This helps make the selection more visible in Raven.}

\item{f_buff}{The frequency buffer (in Hz) that is added to the top and bottom of each selection. This helps make the selection more visible in Raven.}
}
\description{
Generate a Raven selection table from LFDCS autodetections
}
\details{
The Low-Frequency Detection and Classification System (\href{https://www.whoi.edu/cms/files/JASMAN12952889_85804.pdf}{LFDCS}) automatically detects
marine mammal vocalizations in large datasets. This script converts a csv file of
LFDCS autodetections (extracted from raw detector output using `export_autodetections`)
into a text file that can be imported into Raven Pro as a selection table.
}
\examples{
# export all autodetections
lfdcs_to_raven(
 lfdcs_file = 'autodetections.csv',
 raven_file = 'selections.txt',
 audio_start_time = '2017-08-01 17:44:27',
)

# export high quality right whale autodetections
lfdcs_to_raven(
 lfdcs_file = 'autodetections.csv',
 raven_file = 'selections.txt',
 audio_start_time = '2017-08-01 17:44:27',
 call_types = c(5,6,7,8,9),
 max_mdist = 3
)

}
\author{
Hansen Johnson (\email{hansen.johnson@dal.ca})
}
