% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp_dmon.R
\name{timestamp_dmon}
\alias{timestamp_dmon}
\title{Add a timestamp DMON wav files}
\usage{
timestamp_dmon(dmon_dir, prefix = "", suffix = "")
}
\arguments{
\item{dmon_dir}{The directory containing dmon wav and xml files}

\item{prefix}{String to be added to new filename before timestamp}

\item{suffix}{String to be added to new filename after timestamp}
}
\description{
Add a timestamp DMON wav files
}
\details{
Digital acoustic monitoring instruments (DMON) hydrophones log audio data in a `wav` file
and metadata in a `xml` file. This program simply renames the wav file to include the file start time, which
is useful for viewing the data in other sound analysis software (such as Raven Pro).
This is achieved by extracting the start time of the wav file from the associated xml header file.
The original and new timestamps will be saved in an R data file (`dmon_filenames.rds`) in the dmon data directory.
See \code{undo_timestamp_dmon()} for how to convert back to the original timestamps.
NOTE - this will only work for continuous DMON data (i.e., *not* duty cycled).
}
\seealso{
\code{\link{undo_timestamp_dmon}}
}
\author{
Hansen Johnson (\email{hansen.johnson@dal.ca})
}
