% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp_dmon.R
\name{undo_timestamp_dmon}
\alias{undo_timestamp_dmon}
\title{Revert DMON wav files to original file names}
\usage{
undo_timestamp_dmon(timestamp_file)
}
\arguments{
\item{timestamp_file}{The `dmon_filenames.rds` generated by \code{timestamp_dmon()}
containing dmon wav and xml files}
}
\description{
Revert DMON wav files to original file names
}
\seealso{
\code{\link{timestamp_dmon}}
}
\author{
Hansen Johnson (\email{hansen.johnson@dal.ca})
}
